deslop(phase-4): shrink MCP/CLI surface and quarantine compat path
Align search default to hybrid across broker, CLI, schemas, and docs;
demote OpenClaw aliases in listings while keeping wire names; remove dead
prepareEmbeddingsBatch and waxFlush residue; document SearchStrategy/
ResultReranker/Memory.Error as extension points; isolate WaxMCPTools
compat overload into a separate test-oriented file.
