+++ Resources/skills/public/wax-mcp/references/project-rules.md
@@ -10,7 +10,8 @@ Workflow rules:
 - At session start, call `handoff_latest` first to load prior context, then call `session_start` once and keep the returned `session_id`.
 - Use `remember` to store decisions, discoveries, and short factual notes. If the memory is session-scoped, pass `session_id` as a top-level argument. Do not put `session_id` inside `metadata`.
 - Use `recall` for assembled context and `search` for raw ranked hits.
-- Prefer `mode: "hybrid"` when semantic retrieval helps. Use `mode: "text"` when I want a fast or deterministic lexical lookup.
+- `search` / `memory_search` default to `mode: "hybrid"` when omitted. Use `mode: "text"` for fast or deterministic lexical lookup.
+- Prefer `remember` / `memory_promote`. OpenClaw wire aliases `memory_append` / `promote` remain for compatibility.
 - Do not manage `SESSION_STORE`, `--store-path`, or `flush` in normal agent flows. The broker owns long-term memory and virtual session stores.
 - Use `handoff` near the end of the session with `content`, optional `project`, and `pending_tasks`, then call `session_end`.
 - Use `corpus_search` only when you need cross-session retrieval across broker-managed session history with provenance metadata.
